+++ app/models/grid.ts
@@ -3,7 +3,7 @@ import { LayerElements } from "./createLayers";
 
 export default class Grid {
     entities: Array<Entity>;
-    grid: Entity[][];
+    body: Entity[][];
     w: number;
     h: number;
     trim_x:number;
@@ -19,7 +19,7 @@ export default class Grid {
         this.PADDING_RIGHT = layer.elements_padding_right;
         this.PADDING_TOP = layer.elements_padding_top;
         this.popupAvailable = layer.popupAvailable;
-        this.grid = [];
+        this.body = [];
         this.define();
     }
 
@@ -30,19 +30,19 @@ export default class Grid {
             _this.h = entity.height + _this.PADDING_TOP;
             const index_x = Math.floor(entity.x / _this.w) % _this.w;
             const index_y = Math.floor(entity.y / _this.h) % _this.h;
-            if (!_this.grid[index_x]) {
-                _this.grid[index_x] = [];
+            if (!_this.body[index_x]) {
+                _this.body[index_x] = [];
             }
-            _this.grid[index_x][index_y] = entity;
+            _this.body[index_x][index_y] = entity;
         })
     }
 
     getEntity(x:number, y:number):Entity | undefined{
         const index_x = Math.floor(x / this.w) % this.w;
         const index_y = (Math.floor(y / this.h) % this.h);
-        if(this.grid[index_x]){
-            if(this.grid[index_x][index_y]){
-                const entity = this.grid[index_x][index_y];
+        if(this.body[index_x]){
+            if(this.body[index_x][index_y]){
+                const entity = this.body[index_x][index_y];
                 if(entity.checkCoord(x, y)){
                     return entity;
                 }
@@ -52,12 +52,12 @@ export default class Grid {
 
     debug(ctx:CanvasRenderingContext2D){
         const _this = this;
-        this.grid.forEach((col, colindex) => {
+        this.body.forEach((col, colindex) => {
             col.forEach((entity, rowindex) => {
                 let x = colindex * _this.w;
                 let y = rowindex * _this.h;
                 ctx.beginPath()
-                ctx.strokeStyle = _this.color;//"green";
+                ctx.strokeStyle = _this.color;
                 ctx.rect(x, y, _this.w, _this.h)
                 ctx.stroke();
                 ctx.closePath();
